+++ frontend/src/products/mark-tini/MarkTiniApp.tsx
@@ -1,0 +1,5 @@
+import App from '../../App';
+
+export default function MarkTiniApp() {
+  return <App />;
+}